250605AI以外完成/requirements.txt
flask==3.0.0
requests==2.32.3
python-dotenv==1.0.0
gunicorn==21.2.0
